data-viewer-snowflake/requirements.txt
pandas
streamlit
plotly
pyvis
scipy
networkx
snowflake-connector-python
snowflake-snowpark-python==1.10.0